// src/interfaces/common/product-purchase.interface.ts
export interface ProductPurchase {
  _id?: string;
  product?: any;
  previousQuantity?: number;
  shop?: any;
  salesman?: any;
  createTime?: any;
  vendor?: any;
  note?: any;
  imei?: any;
  updatedQuantity?: number;
  createdAtString?: string;
  updatedAtString?: string;
  month?: number;
  year?: number;
}